pyproject.toml
[tool.poetry]
name = "speach-separation"
version = "0.1.0"
description = ""
authors = ["ayk24 <aykueym@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.1.0"
torchaudio = "^2.1.0"
speechbrain = "^0.5.15"
onnxruntime = "^1.16.1"
pyannote-audio = {path = "pyannote-audio"}
polars = "^0.19.9"
pydub = "^0.25.1"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.10.0"
mypy = "^1.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
show_error_context = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.black]
target-version = ['py39']
line-length = 128

[tool.isort]
line_length = 128